location-model: split to separate file
